Extend `NotebookMultiCellAction` and move into contrib
